Discord-bot/requirements.txt
web3
requests
millify
python-dotenv
discord